Air Efficiency, Death and Fade Out
Now the game contains the mechanic of air efficiency which limits the player's stance inside the water, with a base of 45 seconds the player is available to swim around until drowning for which a black screen and a blur effect appear when there's less than 3 seconds to drown.

Also some minor tweaks to the ocean and buoyancy have been made.