Decoys, Items and Landmarks
Now landmarks can be created by using the decoy item which permits the saving of a location in world space and inside the radar. Also few optimizations regarding items got made and now a proper system of the UseSlotItem function has been made which allows any piece of code to be pasted in there to create the item's utility upon usage. Likewise, some structures and enumerations got changed due to optimization and expansion in terms of functionality.