Vehicles 101
Now vehicles have been added to the game, to posses the submarine pawn press P and the camera will instantly switch. The controls consist of physics' torques to maneuver the submarine, the keys for the upwards axis are W/S, for the right axis A/D and for rolling the vehicle Q/E. There isn't a distinction between underwater and above water yet, also the original character isn't attached to the vehicle either. The vehicle is able to take collision damage based on linear acceleration and it has an event included to destroy the vehicle whenever health hits a critical state. Everything added is under the folder of Vehicles at Visuals.