Nestcrab Implementation
Changed the references we were using for the Nestcrab. It simplified a lot of assets and uses solely a skeletal mesh connected to materials for pincers, body and legs. Overall the creature folder should be easier to read. Likewise, BP_Nestcrab has been updated to fit with the implementation as well. The Nestcrab is now ready for adding animation sets.