Nestcrab Movement
The nestcrab is now set up with animations of movement and being idle. Likewise, changed the way we're reading animation blueprints, animation files are now located besides the creature models. Added possesion functionality as well to test movement, press O on the player character to toggle between the Nestcrab and the player.