Base AI Behavior / Footsteps System
Base AI is now set up with a behavior tree for BaseCreature with general functions of finding location, fixing rotation and movement. Also, the structure of different sounds being played by footsteps is set up at the animation blueprint for the Nestcrab.  However, it requires connections to events made through the notify system in animation sequences and to the assets played depending on the surface type.