Upgrade to 5.1
Finally we're on Unreal Engine 5! It's been quite a odyssey trying to upgrade the engine version but it finally compiles. There's a lot of performance improvement out of the box. However, the shaders that use planar reflection need to be adjusted. Likewise, the ocean manager will require a revamp to compile properly in the new version. Nevertheless, it's great to finally have the project on 5.1. The shoal manager and navigation system will be upgraded as soon as they're available for this engine version.